+++ src/main/java/de/maxhenkel/voicechat/voice/common/PlayerState.java
@@ -51,10 +51,16 @@ public class PlayerState {
     }
 
     /**
+     * Empty strings will be treated as null
+     *
      * @param group the group name (Max 16 characters)
      */
     public void setGroup(@Nullable String group) {
-        this.group = group;
+        if (group == null || group.isEmpty()) {
+            this.group = null;
+        } else {
+            this.group = group;
+        }
     }
 
     public boolean hasGroup() {